Title: 합과 평균, Time: 6ms, Memory: 0MB, Status: Wrong Answer - CodeTree

+++ 240915/합과 평균/sum-and-avg.cpp
@@ -3,8 +3,8 @@ using namespace std;
 
 int main() {
     // 여기에 코드를 작성해주세요.
-    int a, b;
+    double a, b;
     cin >> a >> b;
-    cout << (a+b)  << " " << (double)(a+b) / 2 << endl;
+    cout << (a+b)  << " " << (a+b) / 2 << endl;
     return 0;
 }
